src/paper_eval/utils/config.py
from __future__ import annotations
"""Central config helper.
Loads .env vars and exposes constants.
"""

import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-3.5-turbo-0125")
