requirements.txt
PyMuPDF==1.23.22
python-dotenv>=1.0.0
tqdm>=4.66.0
openai>=1.14.0
pytest>=8.0.0
typer[all]>=0.9.0
streamlit>=1.33.0
